% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Danecsv2.R
\docType{data}
\name{Danecsv2}
\alias{Danecsv2}
\title{Dane wlaczone do pakietu}
\format{
Dane zawieraja
\describe{
 \item{Category}{Kategoria Pojazdu}
 \item{Fuel}{Rodzaj Paliwa}
 \item{Segment}{Segment pojazdu}
 \item{Euro Standard}{Norma dopuszczalnych emisji spalin}
 \item{Technology}{Technologia silnika}
 \item{Pollutant}{Rodzaj zanieczyszczenia}
 \item{Mode}{tryb}
 \item{Road Slope}{nachylenie drogi}
 \item{Load}{zaladunek}
 \item{Min Speed km/h}{minimalna predkosc}
 \item{Max Speed km/h}{maksymalna predkosc}
 \item{Alpha}{Alfa}
 \item{Beta}{Beta}
 \item{Gamma}{Gamma}
 \item{Delta}{Delta}
 \item{Epsilon}{Epsilon}
 \item{Zita}{Zita}
 \item{Hta}{Hta}
 \item{Reduction Factor }{Wspolczynnik redukcji}
 \item{Bio Reduction Factor }{Bio Wspolczynnik redukcji}
 \item{50}{kolumna z wartosciami 50}
 \item{EF g/km or ECF MJ/km}{wspolczynnik emisji}
}
}
\usage{
Danecsv2
}
\description{
Pasuja idealnie do wywolywania funkcji kporz_wykres z pakietu newpack2
Sa mniejsze i idealnie dostosowywuja sie do utworzenia wykresu
}
\examples{
\dontrun{
Danecsv2
}
}
\keyword{datasets}
