% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\name{hello2}
\alias{hello2}
\title{Funkcja oblicza srednia z losowych zmiennych}
\usage{
hello2(a = 10, b = 20, d = 2)
}
\arguments{
\item{a}{double}

\item{b}{double}

\item{d}{double}
}
\value{
double
}
\description{
Funkcja oblicza srednia z losowych zmiennych
}
\examples{
hello2(a = 1, b= 500, d =23)
}
