% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kporz_funkcja.R
\name{kporz_funkcja}
\alias{kporz_funkcja}
\title{Obliczenia emisji spalin}
\usage{
kporz_funkcja(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{dane}{dataframe}

\item{kategoria}{character}

\item{euro}{character}

\item{mode}{character}

\item{substancja}{character}
}
\value{

}
\description{
Obliczenia emisji spalin
}
