% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kporz_funkcja_v1.R
\name{kporz_funkcja_v1}
\alias{kporz_funkcja_v1}
\title{kporz_funkcja_v1}
\usage{
kporz_funkcja_v1(
  data = Danecsv,
  Kat = "Passenger Cars",
  Paliwo = NULL,
  Stand = NULL,
  Techn = NULL,
  zaniecz = NULL,
  mod = NULL
)
}
\arguments{
\item{data}{dataframe - Dane zawierajace kolumny z przynajmniej ponizszymi kolumnami}

\item{Kat}{character - kategoria auta}

\item{Paliwo}{character - rodzaj paliwa}

\item{Stand}{character - euro standard}

\item{Techn}{character - rodzaj technologi}

\item{zaniecz}{character - rodzaj zanieczyszczenia}

\item{mod}{character - tryb auta}
}
\value{
dataframe
}
\description{
Funkcja oblicza srednia z losowych zmienn
}
\examples{
\donttest{kporz_funkcja_v1(data = Danecsv,Kat = "Passenger Cars",Paliwo = NULL,
Stand = NULL,Techn = NULL,zaniecz = NULL,mod = "Urban Peak") }
}
