% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kporz_wykres.R
\name{kporz_wykres}
\alias{kporz_wykres}
\title{Funkcja oblicza srednia z losowych zmiennych}
\usage{
kporz_wykres(Danecsv, zm1, zm2, zm3)
}
\arguments{
\item{Danecsv}{dataframe}

\item{zm1}{charackter}

\item{zm2}{charackter}

\item{zm3}{charackter}
}
\value{
dataframe
}
\description{
Funkcja oblicza srednia z losowych zmiennych
}
