% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpack2.R
\docType{package}
\name{newpack2}
\alias{newpack2}
\title{Pakiet oceny emisji (emission evaluation)}
\description{
Pakiet sluzy do obliczania emisji zanieczyszczen generowanych przez pojazdy
z silnikami spalinowymi i wyswietlania wyników na wykresach.
}
\details{
Pakiet zawiera dwie funkcje: kporz_funkcja_v1(),kporz_funkcjav1()

kporz_funkcja_v1() filtruje dane wedlug podanych argumentow i oblicza emisje do nich

kporz_wykres() tworzy jeden z czterech dostepnych wykresow

Dane powinny skladac sie przynajmniej z kolumn: Category,Fuel,Euro Standard, Techonolgy, Pollutant,Mode
aby program dzialal poprawnie
}
