% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kporz_funkcja_v1.R
\name{kporz_funkcja_v1}
\alias{kporz_funkcja_v1}
\title{kporz_funkcja_v1}
\arguments{
\item{data}{dataframe}

\item{Kat}{character}

\item{Paliwo}{character}

\item{Stand}{character}

\item{Techn}{character}

\item{zaniecz}{character}

\item{mod}{character}
}
\value{
dataframe
}
\description{
Funkcja oblicza srednia z losowych zmienn
}
\examples{
kporz_funkcja_v1(data = Danecsv,Kat = "Passenger Cars",Paliwo = NULL,Stand = NULL,Techn = NULL,zaniecz = NULL,mod = "Urban Peak")
}
