% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kporz_wykres.R
\name{kporz_wykres}
\alias{kporz_wykres}
\title{kporz_wykres}
\usage{
kporz_wykres(data = Danecsv, zm1, zm2, zm3, zm4, zm5)
}
\arguments{
\item{data}{dataframe}

\item{zm1}{charackter}

\item{zm2}{charackter}

\item{zm3}{charackter}

\item{zm4}{charackter}

\item{zm5}{charackter}
}
\value{
dataframe
}
\description{
Funkcja oblicza srednia z losowych zmiennych
}
\examples{
\donttest{ kporz_wykres(b4,"Gamma","wynik","Hta","tytuĹ‚ wykresu","wykres1") }

}
